requirements.txt
streamlit==1.41.1
requests==2.32.3
transformers==4.47.1
langchain==0.3.13
langchain-community==0.3.13
openai==1.58.1
python-decouple==3.8
urllib3==2.3.0
